move `Attachment` to its own module so it can be reused. Fix `ToXml` for `Graphics`
