+++ cyclonedx-bom/src/models/mod.rs
@@ -19,6 +19,7 @@
 pub mod advisory;
 pub mod annotation;
 pub mod attached_text;
+pub mod attachment;
 pub mod bom;
 pub mod code;
 pub mod component;
